+++ cocos2d/CCTexture2D.m
@@ -316,7 +316,7 @@ return [NSString stringWithFormat:@"<%@ = %p | Name = %i | Dimensions = %ix%i | 
 			
 			NSUInteger neededBytes = (4 - mod ) / (bpp/8);
             
-			CCLOG(@"cocos2d: WARNING converting size=(%d,%d) to size=(%d,%d) due to iOS 5.x memory BUG. See: http://www.cocos2d-iphone.org/forum/topic/31092", POTWide, POTHigh, POTWide + neededBytes, POTHigh );
+            CCLOG(@"cocos2d: WARNING converting size=(%lu,%lu) to size=(%lu,%lu) due to iOS 5.x memory BUG. See: http://www.cocos2d-iphone.org/forum/topic/31092", (unsigned long)POTWide, (unsigned long)POTHigh, (unsigned long)(POTWide + neededBytes), (unsigned long)POTHigh );
 			POTWide += neededBytes;
 		}
 	}
